# CAB230-A3/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Rental Search API",
    "description": "This API provides access to publicly available data on the Australian rental market in 2026. The API endpoints and their usage are described in detail below."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter JWT Bearer token **_only_**"
      }
    }
  },
  "paths": {
...
    "/rentals/{id}": {
      "get": {
        "tags": [
          "Rentals"
        ],
        "description": "Returns an object containing data for the specified rental property. The path parameter {id} is required.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The ID of the rental property",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object containing title, rent, description, propertyType, locality, latitude, longitude, postcode, state, streetAddress, suburb, bathrooms, bedrooms, parkingSpaces, agencyName, amenities, averageRating and numRatings propertiers for the queried rental.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "examples": [
                        "Great location, large family home"
                      ]
                    },
                    "rent": {
                      "type": "integer",
                      "examples": [
                        870
                      ]
                    },
                    "description": {
                      "type": "string",
                      "examples": [
                        "Tucked away in a quiet cul-de-sac in sought-after Eatons Hill, this light-filled home offers relaxed, easy living in a peaceful setting. Elevated to capture breezes year-round, it feels private, airy and welcoming from the moment you arrive. Designed for effortless indoor-outdoor flow, it's perfect for entertaining or unwinding by the pool. Enjoy the convenience of being within walking distance to buses, Eatons Hill Hotel and local shops. A spacious family home in a friendly, well-connected neighbourhood.<br/><br/>Features:<br/>• King-sized master bedroom with spa ensuite and walk-in robe<br/>• Three queen-sized bedrooms with built-in robes<br/>• Three separate internal living areas<br/>• Large dining area plus breakfast bar<br/>• Air conditioning for year-round comfort<br/>• Insulated undercover outdoor entertaining area<br/>• Solar-heated saltwater pool with feature lighting<br/>• Fully fenced, low-maintenance yard<br/>• Generous storage throughout<br/>• Elevated 795sqm block capturing natural light and breezes<br/>• Walking distance to public transport, Eatons Hill Hotel and speciality shops<br/>• Located within highly regarded school catchments and private school bus routes<br/><br/>This property will not last long. Please note that only those that have viewed the property will have their applications processed.<br/><br/>Information contained on any marketing material, website or other portal should not be relied upon and you should make your own enquiries and seek your own independent advice with respect to any property advertised or the information about the property."
                      ]
                    },
                    "propertyType": {
                      "type": "string",
                      "examples": [
                        "house"
                      ]
                    },
                    "locality": {
                      "type": "string",
                      "examples": [
                        "Eatons Hill"
                      ]
                    },
                    "latitude": {
                      "type": "number",
                      "examples": [
                        -27.33943255
                      ]
                    },
                    "longitude": {
                      "type": "number",
                      "examples": [
                        152.94730591
                      ]
                    },
                    "postcode": {
                      "type": "integer",
                      "examples": [
                        4037
                      ]
                    },
                    "state": {
                      "type": "string",
                      "examples": [
                        "Qld"
                      ]
                    },
                    "streetAddress": {
                      "type": "string",
                      "examples": [
                        "23 Valencia Court"
                      ]
                    },
                    "suburb": {
                      "type": "string",
                      "examples": [
                        "Eatons Hill"
                      ]
                    },
                    "bathrooms": {
                      "type": "integer",
                      "examples": [
                        2
                      ]
                    },
                    "bedrooms": {
                      "type": "integer",
                      "examples": [
                        4
                      ]
                    },
                    "parkingSpaces": {
                      "type": "integer",
                      "examples": [
                        2
                      ]
                    },
                    "agencyName": {
                      "type": "string",
                      "examples": [
                        "McGrath Estate Agents Aspley"
                      ]
                    },
                    "amenities": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "examples": [
                        "Garage: 2, Air Conditioning"
                      ]
                    },
                    "averageRating": {
                      "type": [
                        "number",
                        "null"
                      ],
                      "examples": [
                        4.25
                      ]
                    },
                    "numRatings": {
                      "type": "integer",
                      "examples": [
                        4
                      ]
                    },
                    "reviews": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rating": {
                            "type": "integer",
                            "examples": [
                              4
                            ]
                          },
                          "user": {
                            "type": "string",
                            "examples": [
                              "john.doe@example.com"
                            ]
                          },
                          "comment": {
                            "type": "string",
                            "examples": [
                              "Great place to stay."
                            ]
                          },
                          "dateTime": {
                            "type": "string",
                            "examples": [
                              "2026-03-13T21:51:48.752Z"
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid query parameters: yourParamHere. Query parameters are not permitted."
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "The requested rental property could not be found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "No rental exists with this ID."
                      ]
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/ratings": {
      "get": {
        "tags": [
          "Ratings"
        ],
        "description": "Returns a list of ratings made by the current user. This endpoint requires authorisation to use (click the 'Authorize' button and paste in a valid JWT to test this out in Swagger).",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "An array of objects containing rentalId, rating and dateTime properties. The results are limited to 20 per page. An example of one object in the array is shown below.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "rentalId": {
                            "type": "integer",
                            "examples": [
                              1234
                            ]
                          },
                          "rating": {
                            "type": "integer",
                            "examples": [
                              3
                            ]
                          },
                          "comment": {
                            "type": "string",
                            "examples": [
                              "Decent value for money."
                            ]
                          },
                          "dateTime": {
                            "type": "string",
                            "examples": [
                              "2026-03-14T23:47:45.371Z"
                            ]
                          }
                        }
                      },
                      "examples": [
                        [
                          {
                            "rentalId": 1,
                            "rating": 3,
                            "dateTime": "2026-03-11T23:47:18.496Z"
                          },
                          {
                            "rentalId": 17,
                            "rating": 2,
                            "dateTime": "2026-03-10T18:26:26.153Z"
                          },
                          {
                            "rentalId": 100,
                            "rating": 4,
                            "dateTime": "2026-03-13T21:51:48.752Z"
                          },
                          {
                            "rentalId": 1234,
                            "rating": 3,
                            "dateTime": "2026-03-09T17:39:35.365Z"
                          },
                          {
                            "rentalId": 2000,
                            "rating": 5,
                            "dateTime": "2026-03-12T22:23:14.986Z"
                          }
                        ]
                      ]
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "examples": [
                            12
                          ]
                        },
                        "lastPage": {
                          "type": "integer",
                          "examples": [
                            1
                          ]
                        },
                        "prevPage": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "examples": [
                            null
                          ]
                        },
                        "nextPage": {
                          "type": [
                            "integer",
                            "null"
                          ],
                          "examples": [
                            null
                          ]
                        },
                        "perPage": {
                          "type": "integer",
                          "examples": [
                            20
                          ]
                        },
                        "currentPage": {
                          "type": "integer",
                          "examples": [
                            1
                          ]
                        },
                        "from": {
                          "type": "integer",
                          "examples": [
                            0
                          ]
                        },
                        "to": {
                          "type": "integer",
                          "examples": [
                            20
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "boolean",
                      "examples": [
                        true
                      ]
                    },
                    "message": {
                      "type": "string",
                      "examples": [
                        "Invalid page parameter. Must be an integer greater than or equal to 1."
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Click on 'Schema' below to see the possible error responses.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Authorization header ('Bearer token') not found"
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "JWT token has expired"
                          ]
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "error": {
                          "type": "boolean",
                          "examples": [
                            true
                          ]
                        },
                        "message": {
                          "type": "string",
                          "examples": [
                            "Invalid JWT token"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string",
                  "examples": [
                    "Too many requests, please try again later."
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}